+++ src/deepboard/gui/components/__init__.py
@@ -7,3 +7,4 @@ from .split_selector import SplitSelector
 from .stat_line import StatLine
 from .artefact_group import ArtefactGroup
 from .quick_filter import QuickFilter
+from .plotly2fh import plotly2fasthtml_download