+++ CHANGELOG.md
@@ -9,6 +9,7 @@ data.
 - Now can add and modify notes for each run.
 - Default theme for fragment tables
 - Quick filter: Now can filter categorical columns of the Result Table in the GUI
+- Now can download graph data as CSV files from the GUI
 
 ### Fixed bugs
 - Now can change the port and listening address of the GUI server